requirements.txt
scikit-learn
joblib
openai
python-dotenv
crewai